move skipComment logic
