+++ pkg/config/types.go
@@ -30,24 +30,9 @@ type TestDefinition struct {
 	// Internal field - path to the test file (not in YAML)
 	testFilePath string `yaml:"-"`
 	Skipped      bool   `yaml:"skipped,omitempty"`
-	// Parsed from # SKIPPED preamble comments in the test file (not from YAML).
+	// Parsed from # SKIPPED preamble comments in the test file (not from YAML); see loader.go.
 	commentSkipAll     bool     `yaml:"-"`
 	commentSkipTargets []string `yaml:"-"`
-}
-
-func (t *TestDefinition) ShouldSkipForTarget(targetType string) bool {
-	if t.Skipped {
-		return true
-	}
-	if t.commentSkipAll {
-		return true
-	}
-	for _, x := range t.commentSkipTargets {
-		if x == targetType {
-			return true
-		}
-	}
-	return false
 }
 
 // SetTestFilePath sets the test file path
